style: add memory.go
