+++ internal/dispatcher/dispatcher.go
@@ -2,7 +2,6 @@ package dispatcher
 
 import (
 	"context"
-	"sync"
 
 	"github.com/fljdin/dispatch/internal/tasks"
 )
@@ -45,36 +44,4 @@ func (d *Dispatcher) launchWorkers() {
 		}
 		go worker.Start()
 	}
-}
-
-type Memory struct {
-	wgTasks   sync.WaitGroup
-	wgWorkers sync.WaitGroup
-	queue     tasks.Queue
-	results   chan tasks.Result
-}
-
-func (m *Memory) GetStatus(id int) int {
-	return m.queue.GetStatus(id)
-}
-
-func (m *Memory) SetStatus(id int, status int) {
-	m.queue.SetStatus(id, status)
-}
-
-func (m *Memory) AddTask(task tasks.Task) {
-	m.queue.Add(&task)
-	m.wgTasks.Add(1)
-}
-
-func (m *Memory) ForwardTask(task tasks.Task) {
-	m.queue.Add(&task)
-}
-
-func (m *Memory) StartWorker() {
-	m.wgWorkers.Add(1)
-}
-
-func (m *Memory) EndWorker() {
-	m.wgWorkers.Done()
 }
